Moved rewriting to util.uri and pushed it to the service for Oauth
 * util.view.render_url_template is now util.uri.render_uri_template
   * Kept the uri in the name as it often floats about in files alone
 * The OAuthValidatorService is now responsible for fixing the redirect URL
 * The auth view no longer messes with redirects
   * This also prevents you from putting weird stuff in other parts of the URL
   * Previously they would have been rewritten
 * Moved tests about to compensate
