+++ h/util/uri.py
@@ -291,3 +291,22 @@ def _normalize_queryvalue(value):
 def _blacklisted_query_param(s):
     """Return True if the given string matches any BLACKLISTED_QUERY_PARAMS."""
     return any(re.match(patt, s) for patt in BLACKLISTED_QUERY_PARAMS)
+
+
+def render_uri_template(template, values):
+    """
+    Replace placeholders in a URI from the provided values object.
+
+    This function is primarily used in development to support creating
+    absolute links to h or other Hypothesis services which work when h is
+    accessed from the same system (where the h dev server is "localhost:<port>")
+    or a different device (when the h dev server is "machine-name.local:<port>").
+
+    The ``values`` object must have the following attributes:
+
+     * ``scheme`` - The scheme (e.g. 'http')
+     * ``domain`` - The domain name (e.g. 'example.com')
+    """
+    url = template.replace("{current_host}", values.domain)
+    url = url.replace("{current_scheme}", values.scheme)
+    return url
