+++ h/util/view.py
@@ -25,18 +25,3 @@ def json_view(**settings):
     settings.setdefault("accept", "application/json")
     settings.setdefault("renderer", "json")
     return view_config(**settings)
-
-
-def render_url_template(url_template, request):
-    """
-    Replace placeholders in a URL with elements of the current request's URL.
-
-    This function is primarily used in development to support creating
-    absolute links to h or other Hypothesis services which work when h is
-    accessed from the same system (where the h dev server is "localhost:<port>")
-    or a different device (when the h dev server is "machine-name.local:<port>").
-    """
-    url = url_template
-    url = url.replace("{current_host}", request.domain)
-    url = url.replace("{current_scheme}", request.scheme)
-    return url
