+++ h/views/api/auth.py
@@ -12,7 +12,6 @@ from pyramid.view import view_config, view_defaults
 from h import models
 from h.services.oauth_validator import DEFAULT_SCOPES
 from h.util.datetime import utc_iso8601
-from h.util.view import render_url_template
 from h.views.api.config import api_config
 from h.views.api.exceptions import OAuthAuthorizeError, OAuthTokenError
 
@@ -164,8 +163,8 @@ class OAuthAuthorizeController:
                 )
             )
 
-    def _render_web_message_response(self, redirect_uri):
-        redirect_uri = render_url_template(redirect_uri, self.request)
+    @classmethod
+    def _render_web_message_response(cls, redirect_uri):
         location = urlparse(redirect_uri)
         params = parse_qs(location.query)
         origin = "{url.scheme}://{url.netloc}".format(url=location)
